Remove `source`
Name-wise, this conflicts with some default settings in slog. But it's also covered by the existing `@log` and `@logStream` fields from AWS, and is awkward to define exactly where the boundaries should be.

Products may still want to add their own differentiation (e.g. between the various microservices and frontends that make up Sirius), but we won't specify _how_ they should do that.