Add log structure ADR (#142)
* Add log structure ADR

* Update request and source.

Allow additional request fields, rename "service" to "source" and make it optional.

* Clarify that other fields can be added to log messages

* Remove `source`

Name-wise, this conflicts with some default settings in slog. But it's also covered by the existing `@log` and `@logStream` fields from AWS, and is awkward to define exactly where the boundaries should be.

Products may still want to add their own differentiation (e.g. between the various microservices and frontends that make up Sirius), but we won't specify _how_ they should do that.

* `timestamp` -> `time`

Matches slog's defaults, and is less verbose. I originally chose `timestamp` to match with RFC5424, but none of the tools or libraries we use follow that closely and it's better to have something that's easily actionable.

* PR feedback

Add note about minimum logging level being "INFO" in non-dev environments.

Add `service_name` back in, connecting it to the name of the container or lambda.