Update API standards
This work got left a bit in limbo when the API working group folded, and a few things have changed since then. Most notably, the CDDO standards have been massively improved and now cover a lot of areas that we were filling gaps for.

I've bundled a few changes together here:

**Pointing to the CDDO standards as a starting point.** I've not removed any areas where we're now overlapping (e.g. both now explicitly say "Use JSON") because those overlaps are some of the more important points.

**Simplifying REST explanation.** It's covered in the CDDO documentation, and this isn't the best place to provide a full introduction. Again, I've left the bulk of recommendations in as they are a good precis of the important points.

**General content improvements.** Tidying up a few awkwardly worded things, adding clarity, removing passive voice etc.

**General formatting improvements.** Using bold for and only for important points, adding links, removing trailing periods after code blocks.

I've also taken this opportunity to stop calling these "evolving" standards. All of our standards and guidance are constantly—and intentionally—in flux, but these aren't actively being worked on so don't need to be laid out like so.

#major